requirements.txt
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
ecdsa==0.13
future==0.16.0
idna==2.7
pkg-resources==0.0.0
pyasn1==0.4.4
pycparser==2.19
PyNaCl==1.3.0
python-jose==3.0.1
requests==2.20.0
rsa==4.0
six==1.11.0
urllib3==1.24
